+++ modules/front-end/src/app/core/services/permissions.service.ts
@@ -2,7 +2,7 @@ import {Injectable} from "@angular/core";
 import {lastValueFrom} from "rxjs";
 import {IPolicy, IPolicyStatement} from "@features/safe/iam/types/policy";
 import {MemberService} from "@services/member.service";
-import {EffectEnum, ResourceTypeEnum} from "@features/safe/iam/components/policy-editor/types";
+import {EffectEnum, ResourceTypeEnum} from "@shared/policy";
 
 @Injectable({
   providedIn: 'root'
